Optimise the deletion of versioned folders by reducing the number of queries

+++ mdm-core/grails-app/services/uk/ac/ox/softeng/maurodatamapper/core/container/FolderService.groovy
@@ -37,6 +37,7 @@ import uk.ac.ox.softeng.maurodatamapper.core.model.Model
 import uk.ac.ox.softeng.maurodatamapper.core.model.ModelService
 import uk.ac.ox.softeng.maurodatamapper.path.Path
 import uk.ac.ox.softeng.maurodatamapper.path.PathNode
+import uk.ac.ox.softeng.maurodatamapper.security.SecurableResource
 import uk.ac.ox.softeng.maurodatamapper.security.SecurityPolicyManagerService
 import uk.ac.ox.softeng.maurodatamapper.security.User
 import uk.ac.ox.softeng.maurodatamapper.security.UserSecurityPolicyManager
@@ -156,23 +157,35 @@ class FolderService extends ContainerService<Folder> {
         folder?.deleted = true
     }
 
-    void delete(Folder folder, boolean permanent, boolean flush = true) {
+    Set<SecurableResource> delete(Folder folder, boolean permanent, boolean flush = true, boolean rebuildSecurity = true) {
         if (!folder) {
             log.warn('Attempted to delete Folder which doesnt exist')
-            return
+            return []
         }
         if (permanent) {
-            folder.childFolders.each {delete(it, permanent, false)}
-            modelServices.each {it.deleteAllInContainer(folder)}
+            Set<SecurableResource> deletedResources = []
+            folder.childFolders.each {deletedResources.addAll(delete(it, permanent, false, false))}
+            modelServices.each {deletedResources.addAll(it.deleteAllInContainer(folder, false))}
             folder.trackChanges()
             folder.delete(flush: flush)
             if (securityPolicyManagerService) {
-                securityPolicyManagerService.removeSecurityForSecurableResource(folder, null)
+                if(rebuildSecurity) {
+                    securityPolicyManagerService.removeSecurityForSecurableResource(folder, null)
+                    Map<String, List<SecurableResource>> deletionMap = deletedResources.groupBy {it.domainType}
+                    deletionMap.each {String domainType, List<SecurableResource> resourcesOfType ->
+                        securityPolicyManagerService.removeSecurityForSecurableResourceIds(domainType, resourcesOfType.collect {it.id})
+                    }
+                    return []
+                }
             }
+
+            return deletedResources
         } else {
             folder.childFolders.each {delete(it)}
             delete(folder)
+            return []
         }
+
     }
 
     Folder validate(Folder folder) {
@@ -230,8 +243,10 @@ class FolderService extends ContainerService<Folder> {
     List<Folder> findAllWhereDirectParentOfModel(Model model) {
         List<Folder> folders = []
         Folder modelFolder = get(model.folder.id)
-        folders << modelFolder
-        folders.addAll(findAllWhereDirectParentOfContainer(modelFolder))
+        if(modelFolder) {
+            folders << modelFolder
+            folders.addAll(findAllWhereDirectParentOfContainer(modelFolder))
+        }
         folders
     }
